Game Update
Updated the main menu and lobby with a new design proposal.
Updated the python app.py file import, to search in src file, fixed a
called which had the missing parameter "nickname" + added the gamecode
and nickname specification for the jinja2 in lobby.html.
Added the functionnal game creation button + join game button.